+++ ecs.php
@@ -6,10 +6,8 @@ use Symplify\EasyCodingStandard\Config\ECSConfig;
 
 return static function (ECSConfig $config): void {
     $header = <<<HEADER
-This file is part of the JobRouter Client.
-https://github.com/brotkrueml/jobrouter-client
-
-Copyright (c) 2019-2023 Chris Müller
+This file is part of the JobRouter REST Client.
+https://github.com/jobrouter/php-rest-client
 
 For the full copyright and license information, please view the
 LICENSE.txt file that was distributed with this source code.