+++ src/Client/IncidentsClientDecorator.php
@@ -3,21 +3,19 @@
 declare(strict_types=1);
 
 /*
- * This file is part of the JobRouter Client.
- * https://github.com/brotkrueml/jobrouter-client
- *
- * Copyright (c) 2019-2023 Chris Müller
+ * This file is part of the JobRouter REST Client.
+ * https://github.com/jobrouter/php-rest-client
  *
  * For the full copyright and license information, please view the
  * LICENSE.txt file that was distributed with this source code.
  */
 
-namespace Brotkrueml\JobRouterClient\Client;
+namespace JobRouter\AddOn\RestClient\Client;
 
-use Brotkrueml\JobRouterClient\Enumerations\Priority;
-use Brotkrueml\JobRouterClient\Exception\HttpException;
-use Brotkrueml\JobRouterClient\Model\Incident;
-use Brotkrueml\JobRouterClient\Resource\FileInterface;
+use JobRouter\AddOn\RestClient\Enumerations\Priority;
+use JobRouter\AddOn\RestClient\Exception\HttpException;
+use JobRouter\AddOn\RestClient\Model\Incident;
+use JobRouter\AddOn\RestClient\Resource\FileInterface;
 use Psr\Http\Message\ResponseInterface;
 
 final class IncidentsClientDecorator extends ClientDecorator