+++ src/Exception/InvalidResourceException.php
@@ -3,16 +3,14 @@
 declare(strict_types=1);
 
 /*
- * This file is part of the JobRouter Client.
- * https://github.com/brotkrueml/jobrouter-client
- *
- * Copyright (c) 2019-2023 Chris Müller
+ * This file is part of the JobRouter REST Client.
+ * https://github.com/jobrouter/php-rest-client
  *
  * For the full copyright and license information, please view the
  * LICENSE.txt file that was distributed with this source code.
  */
 
-namespace Brotkrueml\JobRouterClient\Exception;
+namespace JobRouter\AddOn\RestClient\Exception;
 
 final class InvalidResourceException extends \RuntimeException implements ExceptionInterface
 {