+++ tests/Unit/Client/RestClientTest.php
@@ -3,26 +3,24 @@
 declare(strict_types=1);
 
 /*
- * This file is part of the JobRouter Client.
- * https://github.com/brotkrueml/jobrouter-client
- *
- * Copyright (c) 2019-2023 Chris Müller
+ * This file is part of the JobRouter REST Client.
+ * https://github.com/jobrouter/php-rest-client
  *
  * For the full copyright and license information, please view the
  * LICENSE.txt file that was distributed with this source code.
  */
 
-namespace Brotkrueml\JobRouterClient\Tests\Unit\Client;
+namespace JobRouter\AddOn\RestClient\Tests\Unit\Client;
 
-use Brotkrueml\JobRouterClient\Client\RestClient;
-use Brotkrueml\JobRouterClient\Configuration\ClientConfiguration;
-use Brotkrueml\JobRouterClient\Configuration\ClientOptions;
-use Brotkrueml\JobRouterClient\Exception\AuthenticationException;
-use Brotkrueml\JobRouterClient\Exception\HttpException;
-use Brotkrueml\JobRouterClient\Exception\RestClientException;
-use Brotkrueml\JobRouterClient\Resource\FileInterface;
 use donatj\MockWebServer\MockWebServer;
 use donatj\MockWebServer\Response;
+use JobRouter\AddOn\RestClient\Client\RestClient;
+use JobRouter\AddOn\RestClient\Configuration\ClientConfiguration;
+use JobRouter\AddOn\RestClient\Configuration\ClientOptions;
+use JobRouter\AddOn\RestClient\Exception\AuthenticationException;
+use JobRouter\AddOn\RestClient\Exception\HttpException;
+use JobRouter\AddOn\RestClient\Exception\RestClientException;
+use JobRouter\AddOn\RestClient\Resource\FileInterface;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;