+++ tests/Unit/Mapper/RouteContentTypeMapperTest.php
@@ -3,18 +3,16 @@
 declare(strict_types=1);
 
 /*
- * This file is part of the JobRouter Client.
- * https://github.com/brotkrueml/jobrouter-client
- *
- * Copyright (c) 2019-2023 Chris Müller
+ * This file is part of the JobRouter REST Client.
+ * https://github.com/jobrouter/php-rest-client
  *
  * For the full copyright and license information, please view the
  * LICENSE.txt file that was distributed with this source code.
  */
 
-namespace Brotkrueml\JobRouterClient\Tests\Unit\Mapper;
+namespace JobRouter\AddOn\RestClient\Tests\Unit\Mapper;
 
-use Brotkrueml\JobRouterClient\Mapper\RouteContentTypeMapper;
+use JobRouter\AddOn\RestClient\Mapper\RouteContentTypeMapper;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;