+++ tests/Unit/Resource/FileStorageTest.php
@@ -3,19 +3,17 @@
 declare(strict_types=1);
 
 /*
- * This file is part of the JobRouter Client.
- * https://github.com/brotkrueml/jobrouter-client
- *
- * Copyright (c) 2019-2023 Chris Müller
+ * This file is part of the JobRouter REST Client.
+ * https://github.com/jobrouter/php-rest-client
  *
  * For the full copyright and license information, please view the
  * LICENSE.txt file that was distributed with this source code.
  */
 
-namespace Brotkrueml\JobRouterClient\Tests\Unit\Resource;
+namespace JobRouter\AddOn\RestClient\Tests\Unit\Resource;
 
-use Brotkrueml\JobRouterClient\Resource\FileInterface;
-use Brotkrueml\JobRouterClient\Resource\FileStorage;
+use JobRouter\AddOn\RestClient\Resource\FileInterface;
+use JobRouter\AddOn\RestClient\Resource\FileStorage;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;